requirements.txt
ollama==0.6.1
requests==2.32.5
python-dotenv==1.2.1

